% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb.R
\name{nb_hello_world}
\alias{nb_hello_world}
\title{Hello World}
\usage{
nb_hello_world(lang = "English")
}
\arguments{
\item{lang}{language (e.g. "English", "Spanish", "French", etc)}
}
\value{
A list with two values: \code{language} and \code{message}
}
\description{
The helloWorld method returns the string 'Hello, World!' in the language
specified by the language parameter. If the server does not recognize the
specified langauge, it returns the string 'I don't speak that language' in
Traditional Chinese.
}
