% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb.R
\name{nb_list_topics}
\alias{nb_list_topics}
\title{Topic Information}
\usage{
nb_list_topics(scope = c("USER", "ORG", "GLOBAL", "ALL"))
}
\arguments{
\item{scope}{"USER", "ORG", "GLOBAL", or "ALL"}
}
\value{
a \code{tibble}
}
\description{
This function returns a \code{tibble} describing the topics that the currently
logged-in user can access
}
