% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb.R
\name{nb_metric_values}
\alias{nb_metric_values}
\title{Metric Values}
\usage{
nb_metric_values(
  topic_id,
  metricSeries = c("TotalBuzz", "TotalBuzzPost", "PositiveSentiment",
    "NegativeSentiment"),
  time = c("All", "Month", "Week", "Day", "Hour", "QuarterHour", "Minute"),
  ...
)
}
\arguments{
\item{topic_id}{the topic id}

\item{metricSeries}{one of "TotalBuzz", "TotalBuzzPost", "TotalReplies",
"TotalReposts", "OriginalPosts", "Impressions", "PositiveSentiment",
"NegativeSentiment", "NeutralSentiment", "NetSentiment", "Passion",
"UniqueAuthor", "StrongEmotion", "WeakEmotion", "EngagementDislikes",
"EngagementLikes", "EngagementComments", "EngagementShares",
"EngagementTotal", "EngagementRatio", and "EngagementViews"}

\item{time}{one of "All", "Month", "Week", "Day", "Hour", "QuarterHour", "Minute"}

\item{...}{additional arguments. See here: https://nb360.netbase.com/Enterprise/Insight_API/methods/03_insightcount_method}
}
\value{
a \code{tibble}
}
\description{
Metric Values
}
\examples{
\dontrun{
output <- nb_metric_values(topic_id = 1417805, time = "Month", metricSeries = "Impressions")
output <- nb_metric_values(topic_id = 1417805, time = "Day", metricSeries = "TotalBuzz", genders = "Female")
output <- nb_metric_values(topic_id = 1423897, time = "Week", metricSeries = c("TotalBuzz", "TotalBuzzPost", "TotalReplies", "TotalReposts", "OriginalPosts", "Impressions", "PositiveSentiment", "NegativeSentiment", "NeutralSentiment", "NetSentiment", "Passion", "UniqueAuthor", "StrongEmotion", "WeakEmotion", "EngagementDislikes", "EngagementLikes", "EngagementComments", "EngagementShares", "EngagementTotal", "EngagementRatio", "EngagementViews"))
}
}
