% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb.R
\name{nb_setup}
\alias{nb_setup}
\title{User authentication}
\usage{
nb_setup(user, password)
}
\arguments{
\item{user}{user name}

\item{password}{password}
}
\value{
an object created by httr::authenticate() which gets stored invisibly in the global environment as \code{.auth}
}
\description{
User authentication
}
