% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb.R
\name{nb_theme_definitions}
\alias{nb_theme_definitions}
\title{Theme definitions}
\usage{
nb_theme_definitions(theme_id)
}
\arguments{
\item{theme_id}{a vector of theme ids (max 100 at a time)}
}
\value{
a \code{tibble} describing how the themes were created
}
\description{
This function provides information on the themes the currently logged-in user
can access.
}
\details{
If you exceed your rate limit, it will wait for 30 seconds before
trying again, if this fails it will wait for 30 seconds before trying again,
if this fails it will wait for 30 seconds before trying again, if this fails
it will wait for 30 seconds before trying again, if this fails it will wait
for 30 seconds before trying again, if this fails it will wait for 30 seconds
before trying again, if this fails it will wait for 30 seconds before trying
again, if this fails it will wait for 30 seconds before trying again, if this
fails it will wait for 30 seconds before trying again, if this fails it will
wait for 30 seconds before trying again, if this fails it will wait for 30
seconds before trying again...
}
