% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb.R
\name{nb_topic_definitions}
\alias{nb_topic_definitions}
\title{Topic definitions}
\usage{
nb_topic_definitions(topic_id)
}
\arguments{
\item{topic_id}{a vector of topic ids (max 100 at a time)}
}
\value{
a \code{tibble} describing how the topics were created
}
\description{
This function provides information on the topics the currently logged-in user
can access.
}
\details{
If you exceed your rate limit, it will wait for 30 seconds before
trying again, if this fails it will wait for 30 seconds before trying again,
if this fails it will wait for 30 seconds before trying again, if this fails
it will wait for 30 seconds before trying again, if this fails it will wait
for 30 seconds before trying again, if this fails it will wait for 30 seconds
before trying again, if this fails it will wait for 30 seconds before trying
again, if this fails it will wait for 30 seconds before trying again, if this
fails it will wait for 30 seconds before trying again, if this fails it will
wait for 30 seconds before trying again, if this fails it will wait for 30
seconds before trying again...
}
