% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netbase-package.R
\docType{package}
\name{netbase-package}
\alias{netbase}
\alias{netbase-package}
\title{netbase: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: andrés castro araújo \email{acastroaraujo@gmail.com}

}
\keyword{internal}
